+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v0.12.1 ()
+# v0.12.1 (2014/02/10)
 ## New featues
 * *Dead code* finder in puppetresources.
 * CPU related facts.
@@ -6,7 +6,7 @@
 ## Bugs fixed
 * Fixed dependencies so that builds with GHC 7.8-rc1 work.
 
-# v0.12.0 (2014/02/7)
+# v0.12.0 (2014/02/07)
 ## New featues
 * Builds against GHC 7.8-rc1.